requirements.txt
absl-py==0.11.0
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
asttokens==2.4.1
async-timeout==4.0.3
attrdict==2.0.1
attrs==23.2.0
bagpy==0.5
bitarray==2.9.2
bitstring==4.1.4
cachetools==5.3.2
catkin-pkg==1.0.0
certifi==2023.11.17
charset-normalizer==3.3.2
contourpy==1.2.0
cycler==0.12.1
decorator==5.1.1
distro==1.9.0
docutils==0.20.1
exceptiongroup==1.2.0
executing==2.0.1
filelock==3.13.1
fonttools==4.47.0
frozenlist==1.4.1
fsspec==2023.12.2
future==0.18.3
gnupg==2.3.1
google-auth==2.26.1
google-auth-oauthlib==1.2.0
grpcio==1.60.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
imageio==2.33.1
importlib-metadata==7.0.1
importlib-resources==6.1.1
iniconfig==2.0.0
ipython==8.18.1
jedi==0.19.1
Jinja2==3.0.3
kinpy==0.4.2
kiwisolver==1.4.5
lazy_loader==0.3
lxml==4.9.4
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
MouseInfo==0.1.3
mpmath==1.3.0
multidict==6.0.4
networkx==3.2.1
oauthlib==3.2.2
openai==1.8.0
opencv-python==4.7.0.72
packaging==23.2
pandas==2.1.4
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
pluggy==1.3.0
prompt-toolkit==3.0.43
protobuf==4.23.4
psutil==5.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
py3rosmsgs==1.18.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyAutoGUI==0.9.54
pybullet==3.2.6
pycryptodomex==3.19.1
pydantic==2.5.3
pydantic_core==2.14.6
pygame==2.5.2
PyGetWindow==0.0.9
Pygments==2.17.2
PyMsgBox==1.0.9
pyobjc-core==10.1
pyobjc-framework-Cocoa==10.1
pyobjc-framework-Quartz==10.1
pyparsing==3.1.1
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
pyserial==3.5
pytest==7.4.4
python-dateutil==2.8.2
pytorch-kinematics==0.6.0
pytweening==1.0.7
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rospkg==1.5.0
rsa==4.9
rubicon-objc==0.4.7
scikit-build==0.17.6
scikit-image==0.22.0
scipy==1.11.4
seaborn==0.13.1
setuptools-scm==8.0.4
six==1.16.0
sniffio==1.3.0
stack-data==0.6.3
sympy==1.12
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tifffile==2023.12.9
tomli==2.0.1
torch==1.8.0
torchvision==0.9.0
tqdm==4.66.1
traitlets==5.14.1
transformations==2020.1.1
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
vtk==9.3.0
wcwidth==0.2.12
Werkzeug==3.0.1
yarl==1.9.4
zipp==3.17.0
numpy==1.26.0